New downloaders
had to rewrite code that draws downlaoders into an operator.
This now allows a cancel button, but can allow more fun if we have ideas on how to use it - it's s full UI code now, not just draw.
