+++ asset_bar_op.py
@@ -230,6 +230,8 @@ BL_UI_Button.mouse_down_right = mouse_down_right
 BL_UI_Button.set_mouse_down_right = set_mouse_down_right
 
 asset_bar_operator = None
+
+
 # BL_UI_Button.handle_event = handle_event
 
 
@@ -288,7 +290,6 @@ def set_thumb_check(element, asset, thumb_type="thumbnail_small"):
 
     if element.get_image_path() == tpath:
         return
-
     element.set_image(tpath)
     element.set_image_colorspace("")
 
@@ -1067,7 +1068,6 @@ class BlenderKitAssetBarOperator(BL_UI_OT_draw_operator):
             self.show_tooltip()
         if self.active_index != search_index:
             self.active_index = search_index
-
             # scene = bpy.context.scene
             # wm = bpy.context.window_manager
             sr = global_vars.DATA["search results"]
@@ -1077,7 +1077,6 @@ class BlenderKitAssetBarOperator(BL_UI_OT_draw_operator):
             # self.tooltip = asset_data['tooltip']
             ui_props = bpy.context.window_manager.blenderkitUI
             ui_props.active_index = search_index  # + self.scroll_offset
-
             set_thumb_check(self.tooltip_image, asset_data, thumb_type="thumbnail")
             get_tooltip_data(asset_data)
             an = asset_data["displayName"]
@@ -1101,7 +1100,6 @@ class BlenderKitAssetBarOperator(BL_UI_OT_draw_operator):
                     self.version_warning.text = f"Made in Blender {asset_data['sourceAppVersion']}! Some features may not work."
             else:
                 self.version_warning.text = ""
-
             authors = global_vars.DATA["bkit authors"]
             a_id = asset_data["author"]["id"]
             if (
@@ -1218,8 +1216,7 @@ class BlenderKitAssetBarOperator(BL_UI_OT_draw_operator):
     def update_progress_bar(self, asset_button, asset_data):
         if asset_data["downloaded"] > 0:
             pb = asset_button.progress_bar
-            ui_scale = bpy.context.preferences.view.ui_scale
-            w = int(self.button_size * ui_scale * asset_data["downloaded"] / 100.0)
+            w = int(self.button_size * asset_data["downloaded"] / 100.0)
             asset_button.progress_bar.width = w
             asset_button.progress_bar.update(pb.x_screen, pb.y_screen)
             asset_button.progress_bar.visible = True
